backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.7.0

# Database
# SQLite built-in

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# AI
groq==0.9.0
httpx==0.27.0

# Voice
edge-tts==6.1.12
gtts==2.5.1
SpeechRecognition==3.10.4

# Files
PyPDF2==3.0.1
python-docx==1.1.2
aiofiles==23.2.1

# Payments
stripe==10.5.0

# Email
# smtplib built-in

# Utils
Pillow==10.3.0
requests==2.32.3